Package directory was renamed, #6
