+++ markdown_tool/formatters/__init__.py
@@ -1,9 +1,9 @@
-from pkg.formatters.simple import SimpleFormatter
-from pkg.formatters.html import HTMLFormatter
+from markdown_tool.formatters.simple import SimpleFormatter
+from markdown_tool.formatters.html import HTMLFormatter
 from .helpers import format_article, get_formatter
 
 try:
-    from pkg.formatters.pdf import PDFFormatter
+    from markdown_tool.formatters.pdf import PDFFormatter
 except ModuleNotFoundError:
     PDFFormatter = None
