--- pkg/transformers/__init__.py
@@ -1,8 +1,0 @@
-from pkg.transformers.md.transformer import ArticleTransformer as MarkdownArticleTransformer
-from pkg.transformers.html.transformer import ArticleTransformer as HTMLArticleTransformer
-
-from .helpers import transform_article
-
-
-TRANSFORMERS = [MarkdownArticleTransformer, HTMLArticleTransformer]
-__all__ = [TRANSFORMERS, transform_article]
